add PolymerRenderStatus class and PolymerDom.deepContains method
