+++ lib/src/common.dart
@@ -1,0 +1,8 @@
+// Copyright (c) 2016, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+library polymer_interop.src.common;
+
+import 'dart:js';
+
+final JsObject PolymerJs = context['Polymer'];
