+++ lib/src/polymer_base.dart
@@ -17,7 +17,6 @@ import 'dart:js' as js;
 import 'package:web_components/web_components.dart'
     show CustomElementProxyMixin;
 import 'convert.dart';
-import 'custom_event_wrapper.dart';
 
 /// A mixin to make it easier to interoperate with Polymer JS elements.
 ///
