+++ test/behavior_proxy_test.dart
@@ -5,12 +5,12 @@
 @HtmlImport('fixtures/behavior.html')
 library polymer_interop.test.convert_test;
 
-import 'dart:js';
 import 'package:polymer_interop/polymer_interop.dart';
 import 'package:web_components/web_components.dart';
 import 'package:smoke/mirrors.dart' as smoke;
 import 'package:test/test.dart';
 
+/// Uses [smoke]
 main() async {
   await initWebComponents();
