+++ test/polymer_dom_test.html
@@ -1,3 +1,8 @@
+<!--
+Copyright (c) 2015, the Dart project authors.  Please see the AUTHORS file
+for details. All rights reserved. Use of this source code is governed by a
+BSD-style license that can be found in the LICENSE file.
+-->
 <!doctype html>
 <html>
   <head>
